Share a single Language enum across UserPreferences and RecipeRequest
The inline enums generated two incompatible Language types, breaking the
controller that copies the request language into profile preferences.
$ref a shared Language schema so codegen emits one enum everywhere.

Also fix the German-locale ProfilePage test to click "Automatisch".
