+++ services/py-recipe-service/client/cooking_assistant_api_client/models/__init__.py
@@ -5,6 +5,7 @@ from .error_response import ErrorResponse
 from .help_request import HelpRequest
 from .help_request_forwarded import HelpRequestForwarded
 from .help_response import HelpResponse
+from .language import Language
 from .recipe import Recipe
 from .recipe_created import RecipeCreated
 from .recipe_ingredient import RecipeIngredient
@@ -12,11 +13,9 @@ from .recipe_input import RecipeInput
 from .recipe_nutrients import RecipeNutrients
 from .recipe_request import RecipeRequest
 from .recipe_request_forwarded import RecipeRequestForwarded
-from .recipe_request_language import RecipeRequestLanguage
 from .recipe_update import RecipeUpdate
 from .user_credentials import UserCredentials
 from .user_preferences import UserPreferences
-from .user_preferences_language import UserPreferencesLanguage
 from .user_profile import UserProfile
 from .user_profile_update import UserProfileUpdate
 
@@ -26,6 +25,7 @@ __all__ = (
     "HelpRequest",
     "HelpRequestForwarded",
     "HelpResponse",
+    "Language",
     "Recipe",
     "RecipeCreated",
     "RecipeIngredient",
@@ -33,11 +33,9 @@ __all__ = (
     "RecipeNutrients",
     "RecipeRequest",
     "RecipeRequestForwarded",
-    "RecipeRequestLanguage",
     "RecipeUpdate",
     "UserCredentials",
     "UserPreferences",
-    "UserPreferencesLanguage",
     "UserProfile",
     "UserProfileUpdate",
 )
