--- services/py-recipe-service/client/cooking_assistant_api_client/models/user_preferences_language.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class UserPreferencesLanguage(str, Enum):
-    DE = "DE"
-    EN = "EN"
-    HU = "HU"
-
-    def __str__(self) -> str:
-        return str(self.value)
