+++ services/spring-api/src/main/kotlin/org/openapitools/model/Language.kt
@@ -13,19 +13,25 @@ import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Size
-import org.openapitools.model.Language
 import java.util.Objects
 
 /**
- *
- * @param prompt
- * @param language Active UI language; generated recipe content is written in it
- */
-data class RecipeRequest(
-	@get:Size(min = 1, max = 4096)
-	@Schema(example = "null", required = true, description = "")
-	@get:JsonProperty("prompt", required = true) val prompt: kotlin.String,
-	@field:Valid
-	@Schema(example = "null", description = "Active UI language; generated recipe content is written in it")
-	@get:JsonProperty("language") val language: Language? = null,
-)
+* Supported UI and AI-content language as an ISO 639-1 code
+* Values: EN,DE,HU
+*/
+enum class Language(
+	@get:JsonValue val value: kotlin.String,
+) {
+	EN("EN"),
+	DE("DE"),
+	HU("HU"),
+	;
+
+	companion object {
+		@JvmStatic
+		@JsonCreator
+		fun forValue(value: kotlin.String): Language =
+			values().firstOrNull { it -> it.value == value }
+				?: throw IllegalArgumentException("Unexpected value '$value' for enum 'Language'")
+	}
+}
