+++ services/spring-api/src/main/kotlin/org/openapitools/model/UserPreferences.kt
@@ -13,6 +13,7 @@ import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Size
+import org.openapitools.model.Language
 import java.util.Objects
 
 /**
@@ -23,33 +24,13 @@ import java.util.Objects
  * @param aboutMe Free-form user context provided to the AI
  */
 data class UserPreferences(
+	@field:Valid
 	@Schema(example = "null", description = "Preferred UI and AI-content language as an ISO 639-1 code")
-	@get:JsonProperty("language") val language: UserPreferences.Language? = null,
+	@get:JsonProperty("language") val language: Language? = null,
 	@Schema(example = "null", description = "Dietary restriction or style (e.g. vegan, keto)")
 	@get:JsonProperty("diet") val diet: kotlin.collections.List<kotlin.String>? = null,
 	@Schema(example = "null", description = "List of ingredients the user is allergic to")
 	@get:JsonProperty("allergies") val allergies: kotlin.collections.List<kotlin.String>? = null,
 	@Schema(example = "null", description = "Free-form user context provided to the AI")
 	@get:JsonProperty("aboutMe") val aboutMe: kotlin.collections.List<kotlin.String>? = null,
-) {
-	/**
-	 * Preferred UI and AI-content language as an ISO 639-1 code
-	 * Values: EN,DE,HU
-	 */
-	enum class Language(
-		@get:JsonValue val value: kotlin.String,
-	) {
-		EN("EN"),
-		DE("DE"),
-		HU("HU"),
-		;
-
-		companion object {
-			@JvmStatic
-			@JsonCreator
-			fun forValue(value: kotlin.String): Language =
-				values().firstOrNull { it -> it.value == value }
-					?: throw IllegalArgumentException("Unexpected value '$value' for enum 'Language'")
-		}
-	}
-}
+)
